pyproject.toml
[project]
name = "project2"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12.5"
dependencies = [
    "chardet>=5.2.0",
    "httpx>=0.28.1",
    "openai>=1.57.2",
    "pandas>=2.2.3",
    "python-dotenv>=1.0.1",
    "ruff>=0.8.2",
    "seaborn>=0.13.2",
]
